added class AdapterTruckCalculator

+++ lab_07_Adapter/task_3_1/src/AdapterTypeCToMicroUsb.java
@@ -13,6 +13,6 @@ public class AdapterTypeCToMicroUsb implements MicroUsbCharger {
 
     @Override
     public float getOutputAmperage() {
-        return (float) (typeCCharger.getOutputPower()/5.0f);
+        return typeCCharger.getOutputPower()/5.0f;
     }
 }
